+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "node-red-contrib-myuplink",
-  "version": "1.1.3-myuplink-d",
+  "version": "1.1.3-myuplink-e",
   "description": "A Node-Red Node for collecting data from myUplink.",
   "keywords": [
     "node-red",
@@ -17,10 +17,10 @@
     "name": "zinen"
   },
   "files": [
-    "nibeuplink.html",
+    "myuplink.html",
     "examples/"
   ],
-  "main": "nibeuplink.js",
+  "main": "myuplink.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/zinen/node-red-contrib-nibeuplink/tree/myuplink2024"
@@ -32,7 +32,7 @@
   "node-red": {
     "version": ">=1.0.0",
     "nodes": {
-      "myuplink": "nibeuplink.js"
+      "myuplink": "myuplink.js"
     }
   },
   "engines": {
